Add Tariff.class and enum. Changed UserAccountController. Add new change-log

+++ src/main/java/com/exlab/incubator/repository/PersonalAccountRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PersonalAccountRepository extends JpaRepository<PersonalAccount, Integer> {
+public interface PersonalAccountRepository extends JpaRepository<PersonalAccount, Long> {
 }
